new/requirements.txt
# Core FastAPI stack
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangGraph + LangChain
langgraph>=0.0.30
langgraph-checkpoint-postgres>=0.0.5
langchain>=0.1.20
langchain-core>=0.1.30
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Database (PostgreSQL)
psycopg>=3.1.0
psycopg-binary>=3.1.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0

# Frontend
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0

# Utilities
requests>=2.31.0
psutil>=5.9.0

# Testing
pytest>=8.0.0

